Adaptando componente de tour al simulador de estructuras
