+++ src/pages/simulator/components/atoms/ConsoleComponent.tsx
@@ -177,7 +177,7 @@ export function ConsoleComponent({
 
     return (
         <div
-            id="console"
+            data-tour="console"
             className="flex-1 bg-[#101014] text-white mr-2 rounded-xl font-mono"
         >
             {history.map((cmd, index) => (
