+++ src/shared/tour/HighlightBox.tsx
